+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added 
+
+- remove command for removing notifications
+
 ### Fixed
 
 - Admins no longer get cooldowns on commands